requirements.txt
Django>=4.0,<5.0
gunicorn
psycopg2-binary
requests
python-decouple
whitenoise
